Make extra_cmd_args a dict[str, str]
