+++ conf/common/test/llama.toml
@@ -17,10 +17,6 @@
 name = "llama"
 description = "Llama2 70b"
 test_template_name = "NeMoLauncher"
-# FIXME : ~training.model.position_embedding_type was added in the extra_cmd_args in order to fix a bug from NeMo repository (https://github.com/NVIDIA/NeMo).
-# the commit that should fix this issue in NeMo is : 5b296e8af832c67d361fdfb80a165db3affaf76a.
-# Once the new release of NeMoLauncher includes this commit (check by downloading the corresponding container and look inside /opt for this commit), ~training.model.position_embedding_type should be removed from the extra args
-extra_cmd_args = "~training.model.position_embedding_type +training.model.fsdp=True ~training.model.optim.bucket_cap_mb ~training.model.optim.overlap_grad_sync ~training.model.optim.overlap_param_sync ~training.model.optim.contiguous_grad_buffer training.model.virtual_pipeline_model_parallel_size=null training.model.megatron_amp_O2=False training.model.activations_checkpoint_num_layers=null training.model.gradient_accumulation_fusion=False training.model.use_cpu_initialization=True training.model.optim.name=fused_adam training.model.tokenizer.model=TOKENIZER_MODEL training.exp_manager.create_wandb_logger=False"
 
 [cmd_args]
   [cmd_args.training]
@@ -30,3 +26,22 @@ extra_cmd_args = "~training.model.position_embedding_type +training.model.fsdp=T
     [cmd_args.training.model]
     global_batch_size = "256"
     pipeline_model_parallel_size = "2"
+
+# FIXME : ~training.model.position_embedding_type was added in the extra_cmd_args in order to fix a bug from NeMo repository (https://github.com/NVIDIA/NeMo).
+# the commit that should fix this issue in NeMo is : 5b296e8af832c67d361fdfb80a165db3affaf76a.
+# Once the new release of NeMoLauncher includes this commit (check by downloading the corresponding container and look inside /opt for this commit), ~training.model.position_embedding_type should be removed from the extra args
+[extra_cmd_args]
+"~training.model.position_embedding_type" = ""
+"+training.model.fsdp" = "True"
+"~training.model.optim.bucket_cap_mb" = ""
+"~training.model.optim.overlap_grad_sync" = ""
+"~training.model.optim.overlap_param_sync" = ""
+"~training.model.optim.contiguous_grad_buffer" = ""
+"training.model.virtual_pipeline_model_parallel_size" = "null"
+"training.model.megatron_amp_O2" = "False"
+"training.model.activations_checkpoint_num_layers" = "null"
+"training.model.gradient_accumulation_fusion" = "False"
+"training.model.use_cpu_initialization" = "True"
+"training.model.optim.name" = "fused_adam"
+"training.model.tokenizer.model" = "TOKENIZER_MODEL"
+"training.exp_manager.create_wandb_logger" = "False"
