+++ conf/common/test/nccl_test_reduce_scatter.toml
@@ -17,7 +17,6 @@
 name = "nccl_test_reduce_scatter"
 description = "reduce_scatter"
 test_template_name = "NcclTest"
-extra_cmd_args = "--stepfactor 2"
 
 [cmd_args]
 "subtest_name" = "reduce_scatter_perf_mpi"
@@ -26,6 +25,9 @@ extra_cmd_args = "--stepfactor 2"
 "maxbytes" = "4G"
 "iters" = "100"
 "warmup_iters" = "50"
+
+[extra_cmd_args]
+"--stepfactor" = "2"
 
 [extra_env_vars]
 "NCCL_TEST_SPLIT_MASK" = "0x7"
