+++ conf/release/spcx/l40s/test/l40s-bc-nccl_test_all_gather_worst.toml
@@ -17,7 +17,6 @@
 name = "nccl_test_all_gather_worst"
 description = "all_gather"
 test_template_name = "NcclTest"
-extra_cmd_args = "--stepfactor 2"
 
 [cmd_args]
 subtest_name = "all_gather_perf_mpi"
@@ -27,6 +26,9 @@ stepbytes = "0"
 iters = "60"
 check = "0"
 warmup_iters = "40"
+
+[extra_cmd_args]
+"--stepfactor" = "2"
 
 [extra_env_vars]
 "NCCL_TESTS_SPLIT_MASK" = "0x7"
