+++ conf/release/spcx/l40s/test/l40s-bc-nccl_test_bisection.toml
@@ -17,7 +17,6 @@
 name = "nccl_test_bisection"
 description = "nccl_test_bisection"
 test_template_name = "NcclTest"
-extra_cmd_args = "--stepfactor 2"
 
 [cmd_args]
 subtest_name = "bisection_perf_mpi"
@@ -26,3 +25,6 @@ minbytes = "128"
 maxbytes = "4G"
 iters = "100"
 warmup_iters = "50"
+
+[extra_cmd_args]
+"--stepfactor" = "2"
